Add tests for circular dependencies
